% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition-function.R
\name{composition}
\alias{composition}
\title{Get table of estimated weights of transformation}
\usage{
composition(rsdr_object, dimensions = NULL)
}
\arguments{
\item{rsdr_object}{RSDR object, a list of results and parameters. This is
an output from \code{rsdr()} function.}

\item{dimensions}{Dimension index, a vector of integers for the column
indices of the table, indicating the output dimensions transformed by
\code{rsdr()}. The maximum integer is the minimum number of output
dimensions among many dimensional reduction models from differen resampled
subsets.}
}
\value{
RSDR object, a list of results and parameters. Use \code{plot()} to
visualize weights that are used to transformed all input dimension to each
output dimension.
}
\description{
This function get table of estimated weights to transform the input to
output dimensions by \code{rsdr()} function.
}
\examples{

## Create input example
library(medhist)
data(medhistdata)
ps_remover=extract_nps_mh(medhistdata)

mh_bin_nps=
  medhistdata[ps_remover_train$key,] \%>\%
  `exprs<-`(
    exprs(.) \%>\%
      t() \%>\%
      as.data.frame() \%>\%
      rownames_to_column(var='id') \%>\%
      column_to_rownames(var='id') \%>\%
      t()
  ) \%>\%
  trans_binary(verbose=F)
 
input=
  mh_bin_nps \%>\%
  exprs() \%>\%
  t() \%>\%
  as.data.frame()
 
## Fit dimensional reduction models with resampling
rsdr_bin_nps=rsdr(input,'CV',10,'PCA')

## Get table of estimated weights of transformation
composition(rsdr_bin_nps)
}
\keyword{dimensional}
\keyword{matrix,}
\keyword{reduction}
\keyword{rotated}
\keyword{weights,}
