% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation-function.R
\name{transformation}
\alias{transformation}
\title{Transform dimensions using an RSDR model}
\usage{
transformation(
  tidy_set,
  rsdr_object,
  dimensions = NULL,
  input_dim = NULL,
  top_n = NULL,
  output_dim = NULL,
  verbose = T
)
}
\arguments{
\item{tidy_set}{A TidySet (i.e. ExpressionSet) containing the visits of
subjects in outcome dataset, paid by any payment systems. This TidySet also
accomodates outcome dataset. This is an output of
\code{compile_mh_outcome()} from \code{medhist} package, or manually
compiled using \code{ExpressionSet()} from \code{Biobase} package.}

\item{rsdr_object}{RSDR object, a list of results and parameters. This is
an output from \code{rsdr()} function.}

\item{dimensions}{Dimension index, a vector of integers for the column
indices of the table, indicating the output dimensions transformed by
\code{rsdr()}. The maximum integer is the minimum number of output
dimensions among many dimensional reduction models from differen resampled
subsets.}

\item{input_dim}{Input dimension, a vector of characters containing the
names of input dimensions that are selected to transform into output
dimensions.}

\item{output_dim}{Output dimension, a vector of characters containing the
names of output dimensions that are selected after transformation.}

\item{verbose}{Verbosity, a logical indicating whether progress should be
shown.}
}
\value{
A TidySet (i.e. ExpressionSet) containing the transformed table
accessed using \code{exprs()} function from Biobase package. Composition of
input dimensions as a weight table can be accessed using \code{fData()} from
the same package. RSDR models and proportion of variance explained (PVE)
using \code{input_dim} and \code{output_dim} can be accessed using
\code{preproc()}.
}
\description{
This function transforms input dimensions using a re-sampled dimensional
reduction model fitted by \code{rsdr()}.
}
\examples{

## Create input example
library(medhist)
data(medhistdata)
ps_remover=extract_nps_mh(medhistdata)

mh_bin_nps=
  medhistdata[ps_remover_train$key,] \%>\%
  `exprs<-`(
    exprs(.) \%>\%
      t() \%>\%
      as.data.frame() \%>\%
      rownames_to_column(var='id') \%>\%
      column_to_rownames(var='id') \%>\%
      t()
  ) \%>\%
  trans_binary(verbose=F)
 
input=
  mh_bin_nps \%>\%
  exprs() \%>\%
  t() \%>\%
  as.data.frame()
 
## Fit dimensional reduction models with resampling
rsdr_bin_nps=rsdr(input,'CV',10,'PCA')

## Transform dimensions using an RSDR model
dr_bin_nps=
  mh_bin_nps \%>\%
  transformation(
    rsdr_object=rsdr_bin_nps
    ,input_dim=rownames(mh_bin_nps)
    ,top_n=6
    ,verbose=F
  )
}
\keyword{RSDR}
\keyword{transformer}
